cephadm: change bootstrap version tracking
I have removed the CLI commands associated with adding bootstrap
version and time to the version tracker sqlite table. Instead, I
have opted to directly call config-key command to set the KV store
values and to get the KV store values. This will reduce the likelyhood
that clients may modify the bootstrap information stored since there
won't be CLI commands showing up on their menus. I have also added a
new mechanism for checking the status of the bootstrap version. The
mechanism relies on the variable bootstrap_versioned_stored in the
main CephadmOrchestrator class in Cephadm. This boolean variable is
initially false and is changed to true whenever a bootstrap version
was successfully added to the sqlite table. This mechanism was added
to allow quick check for if bootstrap version is added which will save
on overhead since functions that try to add the bootstrap version
exist in the Cephadm serve loop which is called every 60 seconds.

Signed-off-by: Timothy Q Nguyen <timqn22@gmail.com>